+++ buildSrc/src/main/kotlin/AppConfig.kt
@@ -10,6 +10,6 @@ object AppConfig {
     const val targetSdk = 37
     const val javaVersion = 21
 
-    const val versionName = "1.2.9"
-    const val versionCode = 14
+    const val versionName = "1.3.0"
+    const val versionCode = 15
 }